Implement onlineUsers as a set
The array is accessed a lot to search for a user, which means that it
will be iterated through every time. With a large amount of users, this
will be inefficient. Implementing it as a set will allow instant access
to any entry and will also have fast deletion time.